% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_theme.R
\name{le_theme}
\alias{le_theme}
\title{Define the theme of Layman Economics blog}
\usage{
le_theme(scale = 0.8, rm_x_leg = FALSE, rm_y_leg = FALSE)
}
\arguments{
\item{scale}{Size of legend text and title.}

\item{rm_x_leg}{logical value indicating whether x label be removed from chart}

\item{rm_y_leg}{logical value indicating whether y label be removed from chart}
}
\description{
Sets chart borders, font, title size and positioning for html publications
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  le_theme()
}
