% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leplot_col.R
\name{leplot_col}
\alias{leplot_col}
\title{Leplot_col()}
\usage{
leplot_col(
  a,
  ttl,
  lh_units,
  y_range,
  srce = "Source: ABS",
  leg = NULL,
  leg_pos = c(0.02, 0.9),
  leg_col = 1,
  y2_range = NULL,
  var_order = NULL,
  no_leg = 0,
  rh_units = lh_units,
  nudge_rh_units = 0,
  rhs_var = NULL,
  colours = NULL,
  stack = 1,
  flip = 0,
  no_zero = 0,
  thm = "le_theme"
)
}
\arguments{
\item{a}{Dataframe in long format containing 3 columns: Category, variable and value. Categories are on the x axis
NOTE: if you don't want any labelling on the x-axis, set category to "". Useful if you only have one category, but multiple variables.}

\item{ttl}{Title of the graph}

\item{lh_units}{Units for the LHS axis}

\item{y_range}{Y axis range}

\item{srce}{Defaults to "Source: ABS"}

\item{leg}{Legend entries. Defaults to variable names}

\item{leg_pos}{Positioning of legend in cartesian coordinate format}

\item{leg_col}{The desired number of column of legends. Defaults to 1}

\item{y2_range}{Range for a second y axis}

\item{var_order}{Binary or character vector. Set to 1 or 2 for variables in ascending/descending order. Custom order using character vector. Must be same length as number of categories in data frame.}

\item{no_leg}{Binary. Set to 1 if you want to suppress the legend}

\item{rh_units}{Units for second y axis}

\item{nudge_rh_units}{Positioning of rh axis label;ling can be messy - this nudges the label left/right}

\item{rhs_var}{Variable to be plotted on RHS axis}

\item{colours}{Change the order of the colour palette. Input as numbers of the positions you want used first}

\item{stack}{Variables stacked into one column (1), or plotted along the axis (0). Defaults to stacked}

\item{flip}{Binary. Set to 1 for horizontal bar chart. Defaults to 0.}

\item{no_zero}{Suppresses zero line}

\item{thm}{Theme for the chart. Defaults to le_theme()}
}
\value{
h The graph as a list object (ggplot2)
}
\description{
Draws undated columns graphs
}
\examples{
ctry <- rep(c("AUS", 'US', "UK", "NZ"), 3)
yr <- c(rep("2019", 4),rep("2020", 4), rep("2021", 4))
m <- data.frame(category = ctry, variable = yr, value = runif(12))
leplot_col(m, ttl = 'Tittle', lh_units = "Y-axis Unit", y_range = c(0,1.5,.3), flip = 1, stack = 0, leg_pos = c(0.9, 0.9))

}
