% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_rng.R
\name{x_rng}
\alias{x_rng}
\title{x_rng}
\usage{
x_rng(x, d, FY = 0, bar = 0)
}
\arguments{
\item{x}{A numeric vector with start and end years.}

\item{d}{A dataframe}

\item{FY}{Binary. 1 to set range in fiscal years, 0 in calendar years. Defaults to zero.}

\item{bar}{.}
}
\description{
Sets the x axis range to the start and end of the years specified.
}
\examples{
x <- c(2004, 2016)
econ <- economics_long
(x_rng(x, econ))
}
