% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/le_theme.R
\name{le_theme}
\alias{le_theme}
\title{Define the theme of Layman Economics blog}
\usage{
le_theme(leg_pos, lescale = 1, flip = 0)
}
\arguments{
\item{leg_pos}{Legend position}

\item{lescale}{Size of legend text and title. Plot title is text size * 1}

\item{flip}{Binary to flip margins}
}
\value{

}
\description{
Sets chart borders, font, title size and positioning for html publications
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  le_theme()
}
